[refactor] section 로직 변경

+++ src/main/java/subway/station/StationService.java
@@ -1,16 +1,22 @@
 package subway.station;
 
 import org.springframework.stereotype.Service;
+import subway.exceptions.BadRequestException;
+import subway.section.SectionDao;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
+    private static final String USING_STATION = "해당 지하철역은 현재 노선이나 구간에 사용중입니다.";
 
-    public StationService(StationDao stationDao){
+    private StationDao stationDao;
+    private SectionDao sectionDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao){
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse createStation(StationRequest stationRequest){
@@ -26,6 +32,9 @@ public class StationService {
     }
 
     public void deleteById(Long id){
+        if (sectionDao.findByStationId(id).size() >= 2 ){
+            throw new BadRequestException(USING_STATION);
+        }
         stationDao.deleteById(id);
     }
 }
